requirements.txt
asgiref==3.5.1
certifi==2021.10.8
charset-normalizer==2.0.12
crispy-bootstrap5==0.6
cycler==0.11.0
dj-database-url==1.0.0
Django==4.0.4
django-crispy-forms==1.14.0
django-rest-auth==0.9.5
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
et-xmlfile==1.1.0
fonttools==4.33.2
gunicorn==20.1.0
idna==3.3
kiwisolver==1.4.2
matplotlib==3.5.1
mysqlclient==2.1.0
numpy==1.22.3
openpyxl==3.0.9
packaging==21.3
pandas==1.4.2
Pillow==9.1.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycryptodome==3.14.1
PyJWT==2.4.0
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
whitenoise==6.2.0
